generic post URL function

+++ conreq/apps/manage_users/views.py
@@ -1,7 +1,7 @@
 from conreq.utils.apps import generate_context
 from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required, user_passes_test
-from django.http import HttpResponse, JsonResponse
+from django.http import HttpResponse, HttpResponseForbidden, JsonResponse
 from django.template import loader
 from django.views.decorators.cache import cache_page
 
@@ -20,17 +20,20 @@ def manage_users(request):
 @login_required
 @user_passes_test(lambda u: u.is_staff)
 def delete_user(request):
-    try:
-        username = request.GET.get("username", None)
-        user = get_user_model().objects.get(username=username)
+    if request.method == "POST":
+        try:
+            username = request.GET.get("username", None)
+            user = get_user_model().objects.get(username=username)
 
-        # Don't allow staff members to delete eachother
-        if user.is_staff and not request.user.is_superuser:
+            # Don't allow staff members to delete eachother
+            if user.is_staff and not request.user.is_superuser:
+                return JsonResponse({"success": False})
+
+            user.delete()
+
+        except:
             return JsonResponse({"success": False})
 
-        user.delete()
+        return JsonResponse({"success": True})
 
-    except:
-        return JsonResponse({"success": False})
-
-    return JsonResponse({"success": True})
+    return HttpResponseForbidden()
